% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact.R
\docType{data}
\name{EORTC10994}
\alias{EORTC10994}
\title{EORTC10994 dataset}
\format{A data frame with 125 rows and 7 variables}
\description{
A dataset containing treatment, response and covariate information for 125 subjects 
with breast cancer. See 'Details' for the variables in this dataset.
}
\details{
\itemize{
  \item ID. Subject identifier 
  \item Treatment. Treatment received. '0' for control and '1' for experimental
  \item Response. Binary response to treatment - '0' for 'non-responder' and '1' for 'responder'
  \item Age. Age (in years) at diagnosis
  \item TumorSize. size of tumor
  \item Node. Node positive (Yes) or node negative (No)
  \item ERBB2Log2. log2 transformed ERBB2 levels
  }
}
\keyword{datasets}

