% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact.R
\docType{data}
\name{GSE10846}
\alias{GSE10846}
\title{GSE10846 dataset}
\format{A data frame with 412 rows and 1003 columns}
\description{
A dataset containing the survival, treatment and gene expression data for 
412 patients with diffuse large B cell lymphoma and treated with CHOP or 
CHOP+Rituximab
}
\details{
\itemize{
  \item time. Survival time in months
  \item status. Censoring status. '0' for censored, '1' for died
  \item Treatment. Treatment received. '0' for control (CHOP) and '1' 
  for new (CHOP+Rituximab)
  \item Columns 4-1003 are gene expression values (normalized using
   the MAS 5.0 software and log2 transformed) of the first 1000 genes  
   highest variance from the original dataset from GEO
  \item The row names are the array names
  }
}
\keyword{datasets}

