% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact.R
\docType{data}
\name{prostateCancer}
\alias{prostateCancer}
\title{Prostate cancer dataset}
\format{A data frame with 485 rows and 9 variables}
\description{
A dataset containing survival information for 485 subjects with prostate cancer. 
See 'Details' for the variables in this dataset.
}
\details{
\itemize{
  \item ID. Subject identifier 
  \item Treatment. Treatment received. '0' for control and '1' for new
  \item time. Survival time in months
  \item status. Censoring status. '0' for censored, '1' for died
  \item age. Age (in years) at diagnosis
  \item pf. Performance status (Normal Activity or Limited Activity)
  \item sz. Size of the primary tumor (cm2)
  \item sg. Index of a combination of tumor stage and histologic grade 
  \item ap. Serum phosphatic acid phosphatase levels
  }
}
\keyword{datasets}

